Mod doker files.
Organized Docker related matters. After starting the backend for the first time,
you need to run go mod tiny to install github.com/cespare/reflex@latest.
First, the DB starts, the backend starts, and finally the frontend starts, performing a health check and controlling it.
It's annoying that go dependencies are downloaded every time the backend starts, so I bind /go in compose and mount it.
On branch main
Changes to be committed:
	modified:   .gitignore
	modified:   trial-env/backend/Dockerfile
	modified:   trial-env/compose.yaml
	modified:   trial-env/initdb/30_init.sql